requirements.txt
streamlit
langchain
langchain-openai
langchain-community
openai
pyyaml
pandas
pillow
pathlib
plotly